requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
openai==1.12.0
httpx>=0.24.1,<1.0.0
pydub==0.25.1
python-dotenv==1.0.1
requests==2.31.0
aiofiles==23.2.1
pydantic>=2.0.0,<3.0.0
typing-extensions>=4.8.0
annotated-types>=0.5.0
anyio>=3.7.1
idna>=2.8
sniffio>=1.1
starlette>=0.36.3
distro>=1.7.0
ffmpeg-python==0.2.0 